Stephen Hay mode,
Stephen Hay mode is "start with the small screen, expand until it looks like shit. Time to insert a breakpoint!"

In Hay! mode, the viewport shrinks to the lowest default value and then slowly expands the browser. Click the Hay! button again to stop the resizing dead in its tracks.